requirements.txt
amqp==5.3.1
asgiref==3.9.2
async-timeout==5.0.1
billiard==4.2.2
celery==5.3.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.1
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.57.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.4
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
geographiclib==2.1
geopy==2.4.0
gunicorn==21.2.0
idna==3.10
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
Pillow==10.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.7
pycparser==2.23
PyJWT==2.10.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
qrcode==7.4.2
redis==5.0.1
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.6.0
